refactor(api): split auth grouped use-case files into folder-per-use-case
auth 모듈이 22개 use-case 클래스를 4개 복수형 파일(authentication/oauth/password/
account.use-cases.ts)에 묶어 todo 표준(폴더당 1 use-case 파일, SRP)에서 이탈하던 것을
기계적으로 분할한다. 행동 변화 0 — 각 클래스는 workflow로 위임하는 thin 어댑터 그대로다.

- 쓰기 use-case 18종 → application/use-cases/<kebab>/<kebab>.use-case.ts
- 읽기 4종(GetCurrentUser·ListLinkedAccounts·ListActiveSessions·GetOAuthRedirectUri)
  → application/queries/<kebab>/<kebab>.use-case.ts (todo 표준의 read/write 분리 미러)
- use-cases/index.ts 재작성 + queries/index.ts 신설. 소비처(facade 3종·module·specs)는
  쿼리를 ../queries에서 import하도록 갱신.
- workflows/·services/ 내부 오케스트레이션 계층은 유지(Phase 7에서 승인된 내부 패턴으로
  문서화). DI 토큰(클래스 토큰) 변화 없음.

클라이언트 계약 무변경 (openapi 스냅샷 diff 0, auth e2e 85 통과).
