+++ apps/api/src/auth/application/queries/get-oauth-redirect-uri/get-oauth-redirect-uri.use-case.ts
@@ -1,0 +1,10 @@
+import { Injectable } from "@nestjs/common";
+import { OAuthWorkflow } from "../../workflows/oauth.workflow";
+
+@Injectable()
+export class GetOAuthRedirectUriQuery {
+	constructor(private readonly workflow: OAuthWorkflow) {}
+	execute(state: string): ReturnType<OAuthWorkflow["getRedirectUriByState"]> {
+		return this.workflow.getRedirectUriByState(state);
+	}
+}
