+++ apps/api/src/auth/application/use-cases/index.ts
@@ -1,6 +1,24 @@
-export * from "./account/account.use-cases";
-export * from "./authentication/authentication.use-cases";
+export * from "./change-password/change-password.use-case";
+export * from "./complete-oauth-authorization/complete-oauth-authorization.use-case";
+export * from "./delete-account/delete-account.use-case";
+export * from "./exchange-oauth-code/exchange-oauth-code.use-case";
 export * from "./issue-login/issue-login.use-case";
-export * from "./oauth/oauth.use-cases";
-export * from "./password/password.use-cases";
+export * from "./link-oauth-account/link-oauth-account.use-case";
+export * from "./link-oauth-account-with-code/link-oauth-account-with-code.use-case";
+export * from "./login-with-oauth-token/login-with-oauth-token.use-case";
+export * from "./login-with-password/login-with-password.use-case";
+export * from "./logout/logout.use-case";
+export * from "./logout-all/logout-all.use-case";
 export * from "./provision-user/provision-user.use-case";
+export * from "./refresh-tokens/refresh-tokens.use-case";
+export * from "./register/register.use-case";
+export * from "./request-password-reset/request-password-reset.use-case";
+export * from "./request-password-setup-code/request-password-setup-code.use-case";
+export * from "./resend-verification/resend-verification.use-case";
+export * from "./reset-password/reset-password.use-case";
+export * from "./revoke-session/revoke-session.use-case";
+export * from "./set-password/set-password.use-case";
+export * from "./start-oauth-authorization/start-oauth-authorization.use-case";
+export * from "./unlink-oauth-account/unlink-oauth-account.use-case";
+export * from "./update-profile/update-profile.use-case";
+export * from "./verify-email/verify-email.use-case";
