+++ apps/api/src/auth/application/use-cases/request-password-reset/request-password-reset.use-case.ts
@@ -1,0 +1,13 @@
+import { Injectable } from "@nestjs/common";
+import { PasswordWorkflow } from "../../workflows/password.workflow";
+
+@Injectable()
+export class RequestPasswordResetUseCase {
+	constructor(private readonly workflow: PasswordWorkflow) {}
+	execute(
+		email: Parameters<PasswordWorkflow["forgotPassword"]>[0],
+		metadata?: Parameters<PasswordWorkflow["forgotPassword"]>[1],
+	): ReturnType<PasswordWorkflow["forgotPassword"]> {
+		return this.workflow.forgotPassword(email, metadata);
+	}
+}
