+++ apps/api/src/auth/application/use-cases/request-password-setup-code/request-password-setup-code.use-case.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { PasswordWorkflow } from "../../workflows/password.workflow";
+
+@Injectable()
+export class RequestPasswordSetupCodeUseCase {
+	constructor(private readonly workflow: PasswordWorkflow) {}
+	execute(
+		userId: Parameters<PasswordWorkflow["requestPasswordSetupCode"]>[0],
+	): ReturnType<PasswordWorkflow["requestPasswordSetupCode"]> {
+		return this.workflow.requestPasswordSetupCode(userId);
+	}
+}
