+++ apps/api/src/auth/application/use-cases/update-profile/update-profile.use-case.ts
@@ -1,0 +1,13 @@
+import { Injectable } from "@nestjs/common";
+import { CredentialAuthWorkflow } from "../../workflows/credential-auth.workflow";
+
+@Injectable()
+export class UpdateProfileUseCase {
+	constructor(private readonly workflow: CredentialAuthWorkflow) {}
+	execute(
+		userId: Parameters<CredentialAuthWorkflow["updateProfile"]>[0],
+		input: Parameters<CredentialAuthWorkflow["updateProfile"]>[1],
+	): ReturnType<CredentialAuthWorkflow["updateProfile"]> {
+		return this.workflow.updateProfile(userId, input);
+	}
+}
